Refactor header files

+++ src/maths.h
@@ -1,10 +1,9 @@
 // SPDX-FileCopyrightText: © 2025 Phil Armstead <philarmstead@mailbox.org>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-
 #pragma once
 
-#include <stdint.h>
+#include "types.h"
 
 
 static inline uint64_t power(const int base, int exp) {
@@ -15,8 +14,8 @@ static inline uint64_t power(const int base, int exp) {
 
 static inline uint64_t hexBytesToInt(const uint8_t *bytes, const char length) {
 	uint64_t value = 0;
-	for (short i = 0; i < length; ++i) {
-		value += (uint64_t)(bytes[i] * power(256, i));
+	for (uint8_t i = 0; i < length; ++i) {
+		value += bytes[i] * power(256, i);
 	}
 	return value;
 }
